# api/controllers/mobile_nodes.js
var models = require('../database/models.js');
var moment = require('moment');
var config = require('../../config');
var jsonwebtoken = require('jsonwebtoken');



console.log('[Mobile Node API] Ready.');


module.exports = {
    mobile_node_add: mobile_node_add,
    mobile_node_data_regist_mobile_month_add: mobile_node_data_regist_mobile_month_add,
    mobile_node_data_regist_mobile_year_add: mobile_node_data_regist_mobile_year_add,
    mobile_node_data_regist_mobile_add: mobile_node_data_regist_mobile_add,
    mobile_node_get: mobile_node_get,
    mobile_node_data_regist_mobile_month_get: mobile_node_data_regist_mobile_month_get,
    mobile_node_data_regist_mobile_year_get: mobile_node_data_regist_mobile_year_get,
    mobile_node_data_regist_mobile_get: mobile_node_data_regist_mobile_get
};

function mobile_node_add(req, res) {
    console.log('[Mobile Node API] Create Mobile Node.');
    models.Producer.findOne({ Username: req.body.username, Token: req.body.token }, function (err, User) {

        if (User === null) {
            console.log("Invalid session");
            res.status(403).json("Invalid session");
        }
        else {
            models.Mobile_Node.findOne({ Name: req.body.mobile_node_name }, function (err, mobile_node) {
                if (mobile_node == null) {
                    var newRecord = new models.Mobile_Node({
                        Name: req.body.mobile_node_name,
                        Description: req.body.description,
                        Location: req.body.location
                    });
                    console.log(newRecord);

                    newRecord.save(function (err) {
                        if (err) {
                            console.error("Error on saving new record");
                            console.error(err); // log error to Terminal


                        } else {
                            console.log("Created a new record!");
                            //recordCreated(newRecord);
                            res.json({
                                message: 'Object created'
                            })
                        }

                    });
                }
                else {
                    console.log("Mobile Node already registered");
                    res.status(503).json("Object already created");
                }
            }
            );
        }

    });
};

function objectIdWithTimestamp(timestamp) {
    // Convert string date to Date object (otherwise assume timestamp is a date)
    if (typeof (timestamp) == 'string') {
        timestamp = new Date(timestamp);
    }

    // Convert date object to hex seconds since Unix epoch
    var hexSeconds = Math.floor(timestamp / 1000).toString(16);

    // Create an ObjectId with that hex timestamp
    var constructedObjectId = ObjectId(hexSeconds + "0000000000000000");

    return constructedObjectId
}

function mobile_node_get(req, res) {
    models.Mobile_Node.find(function (err, nodes) {
        if (nodes != null) {
            res.status(200).send(nodes);
        }
        else {
            if (err) {
                console.log("No content");
                res.status(204).json("No content");
            }
            else {
                console.log("DB error");
                res.status(500).json("DB Error");
            }
        }
    });
};

function mobile_node_data_regist_mobile_month_add(req, res) {
    console.log('[Mobile Node API] Add Data Month.');
    models.Producer.findOne({ Username: req.body.username, Token: req.body.token }, function (err, User) {

        if (User === null) {
            console.log("Invalid session");
            res.status(403).json("Invalid session");
        }
        else {
            models.Mobile_Node.findOne({ Name: req.body.mobile_node_name }, function (err, mobile_node) {
                if (mobile_node == null) {
                    console.log("Mobile Node does not exist");
                    res.status(503).json("Mobile Node does not exist")
                }
                else {


                    models.Data_Type.findOne({ Name: req.body.data_type }, function (err, data_type) {
                        if (data_type == null) {
                            console.log("Data Type does not exist");
                            res.status(503).json("Data Type does not exist")
                        }

                        else {
                            var newRecord = new models.Data_Regist_Mobile_Month({
                                Data_Type_ID: data_type._id,
                                Mobile_Node_ID: mobile_node._id,
                                Value: req.body.value,
                                Month: req.body.month,
                                Year: req.body.year
                            });
                            console.log(newRecord);

                            newRecord.save(function (err) {
                                if (err) {
                                    console.error("Error on saving new record");
                                    console.error(err); // log error to Terminal


                                } else {
                                    console.log("Created a new record!");
                                    //recordCreated(newRecord);
                                    res.json({
                                        message: 'Object created'
                                    })
                                }
                            });
                        }
                    });
                }
            }
            );
        }
    });
};

function mobile_node_data_regist_mobile_month_get(req, res) {
    models.Mobile_Node.findOne({ Name: req.get("gate_name") }, function (err, node) {
        if (gate == null) {
            console.log("Gate does not exist");
            res.status(503).json("Gate does not exist")
        }
        else {
            //ano primeiro, de seguida mes e depois dia
            if (req.get("date") == null) {
                date_search = "1980/01/01";
            }
            else {
                date_search = req.get("date");
            }
            models.Data_Regist_Mobile_Month.find({ Mobile_Node_ID: node._id, _id: { $gt: objectIdWithTimestamp(date_search) } }, function (err, data_event) {
                if (data_event != null) {
                    res.status(200).send(data_event);

                }
                else {
                    if (err) {
                        console.log("DB error");
                        res.status(500).json("DB Error");
                    } else {
                        console.log("No content");
                        res.status(204).json("No content");
                    }
                }
            });
        }
    });
};

function mobile_node_data_regist_mobile_year_add(req, res) {
    console.log('[Mobile Node API] Add Data Year.');
    models.Producer.findOne({ Username: req.body.username, Token: req.body.token }, function (err, User) {

        if (User === null) {
            console.log("Invalid session");
            res.status(403).json("Invalid session");
        }
        else {
            models.Mobile_Node.findOne({ Name: req.body.mobile_node_name }, function (err, mobile_node) {
                if (mobile_node == null) {
                    console.log("Mobile Node does not exist");
                    res.status(503).json("Mobile Node does not exist")
                }
                else {


                    models.Data_Type.findOne({ Name: req.body.data_type }, function (err, data_type) {
                        if (data_type == null) {
                            console.log("Data Type does not exist");
                            res.status(503).json("Data Type does not exist")
                        }

                        else {
                            var newRecord = new models.Data_Regist_Mobile_Year({
                                Data_Type_ID: data_type._id,
                                Mobile_Node_ID: mobile_node._id,
                                Value: req.body.value,
                                Year: req.body.year
                            });
                            console.log(newRecord);

                            newRecord.save(function (err) {
                                if (err) {
                                    console.error("Error on saving new record");
                                    console.error(err); // log error to Terminal


                                } else {
                                    console.log("Created a new record!");
                                    //recordCreated(newRecord);
                                    res.json({
                                        message: 'Object created'
                                    })
                                }
                            });
                        }
                    });
                }
            }
            );
        }
    });
};

function mobile_node_data_regist_mobile_year_get(req, res) {
    models.Mobile_Node.findOne({ Name: req.get("gate_name") }, function (err, node) {
        if (gate == null) {
            console.log("Gate does not exist");
            res.status(503).json("Gate does not exist")
        }
        else {
            //ano primeiro, de seguida mes e depois dia
            if (req.get("date") == null) {
                date_search = "1980/01/01";
            }
            else {
                date_search = req.get("date");
            }
            models.Data_Regist_Mobile_Year.find({ Mobile_Node_ID: node._id, _id: { $gt: objectIdWithTimestamp(date_search) } }, function (err, data_event) {
                if (data_event != null) {
                    res.status(200).send(data_event);

                }
                else {
                    if (err) {
                        console.log("DB error");
                        res.status(500).json("DB Error");
                    } else {
                        console.log("No content");
                        res.status(204).json("No content");
                    }
                }
            });
        }
    });
};

function mobile_node_data_regist_mobile_add(req, res) {
    console.log('[Mobile Node API] Add Data Year.');
    models.Producer.findOne({ Username: req.body.username, Token: req.body.token }, function (err, User) {

        if (User === null) {
            console.log("Invalid session");
            res.status(403).json("Invalid session")
        }
        else {
            models.Mobile_Node.findOne({ Name: req.body.mobile_node_name }, function (err, mobile_node) {
                if (mobile_node == null) {
                    console.log("Mobile Node does not exist");
                    res.status(503).json("Mobile Node does not exist")
                }
                else {


                    models.Data_Type.findOne({ Name: req.body.data_type }, function (err, data_type) {
                        if (data_type == null) {
                            console.log("Data Type does not exist");
                            res.status(503).json("Data Type does not exist")
                        }

                        else {
                            var newRecord = new models.Data_Regist_Mobile({
                                Data_Type_ID: data_type._id,
                                Mobile_Node_ID: mobile_node._id,
                                Value: req.body.value,
                                Date: req.body.date
                            });
                            console.log(newRecord);

                            newRecord.save(function (err) {
                                if (err) {
                                    console.error("Error on saving new record");
                                    console.error(err); // log error to Terminal


                                } else {
                                    console.log("Created a new record!");
                                    //recordCreated(newRecord);
                                    res.json({
                                        message: 'Object created'
                                    })
                                }
                            });
                        }
                    });
                }
            }
            );
        }
    });
};

function mobile_node_data_regist_mobile_get(req, res) {
    models.Mobile_Node.findOne({ Name: req.get("gate_name") }, function (err, node) {
        if (gate == null) {
            console.log("Gate does not exist");
            res.status(503).json("Gate does not exist")
        }
        else {
            //ano primeiro, de seguida mes e depois dia
            if (req.get("date") == null) {
                date_search = "1980/01/01";
            }
            else {
                date_search = req.get("date");
            }
            models.Data_Regist_Mobile.find({ Mobile_Node_ID: node._id, _id: { $gt: objectIdWithTimestamp(date_search) } }, function (err, data_event) {
                if (data_event != null) {
                    res.status(200).send(data_event);

                }
                else {
                    if (err) {
                        console.log("DB error");
                        res.status(500).json("DB Error");
                    } else {
                        console.log("No content");
                        res.status(204).json("No content");
                    }
                }
            });
        }
    });
};