# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.2.1"
  description: "API made for the SHCITY project"
  title: SHCity API
  contact: 
    name: "Alexandre Dias"
    url: "https://www.linkedin.com/in/alexandredias95/"
    email: "alex.m.a.dias95@gmail.com" 
# during dev, should point to your local machine
host: 193.136.127.202:80
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /mobile_node/mobile_node_add: 
    x-swagger-router-controller: mobile_nodes
    post:
      summary: Add a Mobile_node.
      tags:
              - mobile_node
      consumes:
        - application/json
      operationId: mobile_node_add
      parameters:
        - in: body
          name: Mobile_node
          description: The Mobile_node to create.
          schema:
            $ref: "#/definitions/Mobile_node"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
  /mobile_node/mobile_node_data_regist_mobile_month_add: 
    x-swagger-router-controller: mobile_nodes
    post:
      summary: Add a Mobile_node value.
      tags:
              - mobile_node
      consumes:
        - application/json
      operationId: mobile_node_data_regist_mobile_month_add
      parameters:
        - in: body
          name: Mobile_node
          description: The Mobile_node to add the value.
          schema:
            $ref: "#/definitions/Mobile_node_month"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
  /mobile_node/mobile_node_data_regist_mobile_year_add: 
    x-swagger-router-controller: mobile_nodes
    post:
      summary: Add a Mobile_node value.
      tags:
              - mobile_node
      consumes:
        - application/json
      operationId: mobile_node_data_regist_mobile_year_add
      parameters:
        - in: body
          name: Mobile_node
          description: The Mobile_node to add the value.
          schema:
            $ref: "#/definitions/Mobile_node_year"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
  /mobile_node/mobile_node_data_regist_mobile_add: 
    x-swagger-router-controller: mobile_nodes
    post:
      summary: Add a Mobile_node value.
      tags:
              - mobile_node
      consumes:
        - application/json
      operationId: mobile_node_data_regist_mobile_add
      parameters:
        - in: body
          name: Mobile_node
          description: The Mobile_node to add the value.
          schema:
            $ref: "#/definitions/Mobile_node_date"
          required: true
      responses:
        201:
          description: Created
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        503:
          description: Error
          schema:
            $ref: "#/definitions/HelloWorldResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  Mobile_node:
    properties:  
        username:
          type : string
        token:
          type : string
        mobile_node_name:
          type : string
        description:
          type : string 
        location:
          type : string 
  Mobile_node_month:
    properties:  
        username:
          type : string
        token:
          type : string
        mobile_node_name:
          type : string
        data_type:
          type : string 
        value:
          type : string
        month:
          type : string  
        year:
          type : string
  Mobile_node_year:
    properties:  
        username:
          type : string
        token:
          type : string
        mobile_node_name:
          type : string
        data_type:
          type : string 
        value:
          type : string
        year:
          type : string
  Mobile_node_date:
    properties:  
        username:
          type : string
        token:
          type : string
        mobile_node_name:
          type : string
        data_type:
          type : string 
        value:
          type : string
        date:
          type : string
...